fix: move ui components
- Moved components like button, input, table, dropdown in the UI folder
- Updated imports
