+++ src/components/RestoreUI.tsx
@@ -23,8 +23,8 @@ import {
 } from '@heroicons/react/20/solid'
 import { shortenDID } from '@/lib/ui'
 import { Popover, PopoverButton, PopoverPanel } from '@headlessui/react'
-import Button from './Button'
-import Input from './Input'
+import Button from '@/components/ui/Button'
+import Input from '@/components/ui/Input'
 import { Key, useKeychainContext } from '@/contexts/keychain'
 import Keychain from './Keychain'
 import { cidUrl } from '@/lib/storacha'
