+++ src/stories/ui/Button.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import Button from '@/components/Button'
+import Button from '@/components/ui/Button'
 import { ArrowRightIcon, ChevronDownIcon } from '@heroicons/react/24/solid'
 
 const meta = {
