+++ src/stories/ui/Input.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import Input from '@/components/Input'
+import Input from '@/components/ui/Input'
 
 const meta: Meta<typeof Input> = {
   title: 'UI/Input',
